0.0.5
First commit.
0.0.5 is the version number, assuming 0.5 is going to be a working
version of the game in BlueJ, 0.6-0.8 could be balances, or possible
additions to the game, and 1.0 is a working unity prototype, with
further versions on top of that being additions to the unity version (or
maybe a complete rewrite in C# becuase I did some simple math the other
day and C# > Java.